Function docs::routes::about::info

pub async fn info(context: Context<'_>) -> Result<Json<String>, ApiError>

/about
Returns About

About

The about endpoint provides information of the environmental variables set for the instance of safe-client-gateway. This would allow to identify on which commit and version the last deployment happened and to which safe transaction service backend environment the current instance of the gateway is pointing to.

Path

/about

Query parameters

There are no query parameters for this endpoint