Function safe_client_gateway::routes::transactions::details

pub async fn details(
    context: Context<'_>, 
    details_id: String
) -> Result<Json<String>, ApiError>

/v1/transactions/<transaction_id>
Returns TransactionDetails

Transaction Details

The transaction details endpoint provides additional information for a transaction, in much more detail than what the transaction summary endpoint does. It returns a single object crate::models::service::transactions::details::TransactionDetails

Path

/v1/transactions/<transaction_id>

<transaction_id> can be either an id returned by the transaction summary list endpoint or a safe_tx_hash from the Safe Transaction API.

Query paramets

There aren’t any query parameters that can be passed to this endpoint.