Function docs::routes::transactions::propose_transaction

pub async fn propose_transaction<'e>(
    context: Context<'_>, 
    safe_address: String, 
    multisig_transaction_request: Result<Json<MultisigTransactionRequest>, JsonError<'e>>
) -> Result<(), ApiError>

/v1/transactions/<safe_address>/propose
No return value

Transaction Proposal

This endpoint provides a way for submitting transactions of any kind in the format expected by the core services. See the example json to see how to submit a cancellation transaction (you would need to supply a nonce, signature and contractTransactionHash appropriate to the transaction you are submitting)

Path

POST /v1/transactions/<safe_address>/propose

The expected crate::models::service::transactions::requests::MultisigTransactionRequest body for this request, can be found in the sections models

Query parameters

No query parameters available for this endpoint.